# docs/backend-contracts/contract-history/1.3.0-contract-914084874afda2481813a041da4cc01249aa9ea557d9a8bf29baeed4f10e0dc9/openapi.snapshot.yaml
openapi: 3.1.0
info:
  title: BNBU Sports Unified API
  version: 1.3.0-contract
  summary: Unified contract for the sports check-in platform
  description: |
    Stage 21B contract. It does not prove that every registered capability is
    business-enabled or production-ready.
    Business fields come from 02-data-dictionary.md, state transitions from
    03-state-machines.md, rules from 04-business-rules.md, and authorization
    from 05-permission-matrix.md. Foundation authentication/session persistence
    is approved; CourseInvite/Join Capability are contract-and-persistence-design
    only until the Enrollment Gate, while every Export operation is fail-closed
    and V1 persists no ExportJob. Of the 30 client-support operations added in
    Stage 21, authentication, exemption, release-policy, notification,
    push-device, preference, help, and feedback capabilities are approved for
    real local implementation, while unresolved capabilities remain explicit
    routed denials. GPS has persistence and
    application-layer foundations, but all six location HTTP operations remain
    denied. None of these statements proves Staging, iOS-binary, push-delivery,
    GPS-production, or Production-Gate readiness.
servers:
  - url: /api/v1
    description: Version 1 relative base path
security:
  - bearerAuth: []
...
paths:
...
  /class-sections/{classSectionId}/roster-imports:
    get:
      tags: [Official Roster Imports]
      operationId: listRosterImports
      x-access-policy:
        policyId: ROSTER-IMPORT-LIST
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: ROLE_SCOPED
        resourceResolver: ROSTER_CLASS_SECTION_READ_SCOPE
        defaultDeny: true
      summary: List official roster import versions for an authorized class section
      parameters:
        - $ref: '#/components/parameters/ClassSectionId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/RosterImportStatus' }
      responses:
        '200':
          $ref: '#/components/responses/RosterImportListSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    post:
      tags: [Official Roster Imports]
      operationId: createRosterImport
      x-access-policy:
        policyId: ROSTER-IMPORT-CREATE
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: TEACHER_CLASS_SECTION
        resourceResolver: CLASS_SECTION_FROM_PATH
        defaultDeny: true
      summary: Upload a new immutable official roster version
      parameters:
        - $ref: '#/components/parameters/ClassSectionId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            encoding:
              file:
                contentType: text/csv
            schema:
              $ref: '#/components/schemas/CreateRosterImportRequest'
      responses:
        '201':
          $ref: '#/components/responses/RosterImportSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /class-sections/{classSectionId}/roster-imports/current:
    get:
      tags: [Official Roster Imports]
      operationId: getCurrentRosterImport
      x-access-policy:
        policyId: ROSTER-IMPORT-CURRENT-READ
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: ROLE_SCOPED
        resourceResolver: ROSTER_CLASS_SECTION_READ_SCOPE
        defaultDeny: true
      summary: Get the current validated official roster version for an authorized class section
      parameters:
        - $ref: '#/components/parameters/ClassSectionId'
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          $ref: '#/components/responses/RosterImportSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /roster-imports/{rosterImportId}:
    get:
      tags: [Official Roster Imports]
      operationId: getRosterImport
      x-access-policy:
        policyId: ROSTER-IMPORT-READ
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: ROLE_SCOPED
        resourceResolver: ROSTER_IMPORT_READ_SCOPE
        defaultDeny: true
      summary: Get an official roster import and validation summary
      parameters:
        - $ref: '#/components/parameters/RosterImportId'
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          $ref: '#/components/responses/RosterImportSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /roster-imports/{rosterImportId}/rollback:
    post:
      tags: [Official Roster Imports]
      operationId: rollbackRosterImport
      x-access-policy:
        policyId: ROSTER-IMPORT-ROLLBACK
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: TEACHER_CLASS_SECTION
        resourceResolver: ROSTER_IMPORT_FROM_PATH
        defaultDeny: true
      summary: Atomically make a historical validated roster version current
      parameters:
        - $ref: '#/components/parameters/RosterImportId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RollbackRosterImportRequest'
      responses:
        '200':
          $ref: '#/components/responses/RosterImportSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /roster-imports/{rosterImportId}/entries:
    get:
      tags: [Official Roster Imports]
      operationId: listRosterEntries
      x-access-policy:
        policyId: ROSTER-ENTRY-LIST
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: ROLE_SCOPED
        resourceResolver: ROSTER_IMPORT_READ_SCOPE
        defaultDeny: true
      summary: List validated source rows for an authorized import
      parameters:
        - $ref: '#/components/parameters/RosterImportId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Search'
        - name: rowValidationStatus
          in: query
          schema: { $ref: '#/components/schemas/RosterRowValidationStatus' }
      responses:
        '200':
          $ref: '#/components/responses/RosterEntryListSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /roster-imports/{rosterImportId}/align:
    post:
      tags: [Roster Alignment]
      operationId: alignRosterImport
      x-access-policy:
        policyId: ROSTER-IMPORT-ALIGN
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: TEACHER_CLASS_SECTION
        resourceResolver: ROSTER_IMPORT_FROM_PATH
        defaultDeny: true
      summary: Run an immutable roster comparison revision
      parameters:
        - $ref: '#/components/parameters/RosterImportId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunAlignmentRequest'
      responses:
        '202':
          $ref: '#/components/responses/AlignmentRunSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /roster-alignment-results:
    get:
      tags: [Roster Alignment]
      operationId: listRosterAlignmentResults
      x-access-policy:
        policyId: ROSTER-ALIGNMENT-LIST
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: ROLE_SCOPED
        resourceResolver: ROSTER_ALIGNMENT_LIST_SCOPE
        defaultDeny: true
      summary: List authorized immutable classification results
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/ClassSectionFilter'
        - name: rosterImportId
          in: query
          schema: { $ref: '#/components/schemas/OpaqueId' }
        - name: alignmentRunId
          in: query
          schema: { $ref: '#/components/schemas/OpaqueId' }
        - name: currentOnly
          in: query
          schema: { type: boolean, default: true }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/RosterAlignmentStatus' }
        - name: resolutionStatus
          in: query
          schema: { $ref: '#/components/schemas/RosterResolutionStatus' }
      responses:
        '200':
          $ref: '#/components/responses/AlignmentResultListSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /roster-alignment-results/{alignmentResultId}:
    get:
      tags: [Roster Alignment]
      operationId: getRosterAlignmentResult
      x-access-policy:
        policyId: ROSTER-ALIGNMENT-READ
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: ROLE_SCOPED
        resourceResolver: ROSTER_ALIGNMENT_READ_SCOPE
        defaultDeny: true
      summary: Get an authorized alignment result and current resolution projection
      parameters:
        - $ref: '#/components/parameters/AlignmentResultId'
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          $ref: '#/components/responses/AlignmentResultSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /roster-alignment-results/{alignmentResultId}/confirm:
    post:
      tags: [Roster Alignment]
      operationId: confirmRosterAlignmentResult
      x-access-policy:
        policyId: ROSTER-ALIGNMENT-CONFIRM
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: TEACHER_CLASS_SECTION
        resourceResolver: ROSTER_ALIGNMENT_FROM_PATH
        defaultDeny: true
      summary: Confirm a pending exception without changing its classification
      parameters:
        - $ref: '#/components/parameters/AlignmentResultId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        $ref: '#/components/requestBodies/VersionedReasonBody'
      responses:
        '200':
          $ref: '#/components/responses/AlignmentResultSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /roster-alignment-results/{alignmentResultId}/resolve:
    post:
      tags: [Roster Alignment]
      operationId: resolveRosterAlignmentResult
      x-access-policy:
        policyId: ROSTER-ALIGNMENT-RESOLVE
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: TEACHER_CLASS_SECTION
        resourceResolver: ROSTER_ALIGNMENT_FROM_PATH
        defaultDeny: true
      summary: Resolve an exception with traceable evidence
      parameters:
        - $ref: '#/components/parameters/AlignmentResultId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveAlignmentRequest'
      responses:
        '200':
          $ref: '#/components/responses/AlignmentResultSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /roster-alignment-results/{alignmentResultId}/ignore:
    post:
      tags: [Roster Alignment]
      operationId: ignoreRosterAlignmentResult
      x-access-policy:
        policyId: ROSTER-ALIGNMENT-IGNORE
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: TEACHER_CLASS_SECTION
        resourceResolver: ROSTER_ALIGNMENT_FROM_PATH
        defaultDeny: true
      summary: Ignore an allowed exception with a required reason after policy approval
      description: Contract placeholder. Always returns ROSTER_IGNORE_NOT_ALLOWED with no side effects until ADR-057 is accepted.
      x-enabled-by-default: false
      x-business-blocker: ADR-057
      x-default-deny-error: ROSTER_IGNORE_NOT_ALLOWED
      parameters:
        - $ref: '#/components/parameters/AlignmentResultId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        $ref: '#/components/requestBodies/VersionedReasonBody'
      responses:
        '200':
          $ref: '#/components/responses/AlignmentResultSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /roster-alignment-results/{alignmentResultId}/reopen:
    post:
      tags: [Roster Alignment]
      operationId: reopenRosterAlignmentResult
      x-access-policy:
        policyId: ROSTER-ALIGNMENT-REOPEN
        authentication: ACCESS_TOKEN
        allowedRoles: [TEACHER]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: TEACHER_CLASS_SECTION
        resourceResolver: ROSTER_ALIGNMENT_FROM_PATH
        defaultDeny: true
      summary: Reopen a resolved or ignored result
      parameters:
        - $ref: '#/components/parameters/AlignmentResultId'
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        $ref: '#/components/requestBodies/VersionedReasonBody'
      responses:
        '200':
          $ref: '#/components/responses/AlignmentResultSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token carried only in the Authorization header.
    JoinCapability:
      type: apiKey
      in: header
      name: X-Join-Capability
      description: Short-lived one-time capability bound to one ClassSection. The plaintext value is never logged or persisted; no ordinary Access Token is accepted by the Join operation.
  parameters:
    RequestIdHeader:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller correlation ID; the server returns the accepted/generated value in response header and meta.requestId.
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: '^[A-Za-z0-9._:-]+$'
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: Reuse for retries of the same normalized request; reuse with different content returns a conflict.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[\x21-\x7E]+$'
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor bound to caller scope, filters, and sort.
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      required: false
      description: Comma-separated endpoint whitelist; prefix a field with '-' for descending order.
      schema:
        type: string
        maxLength: 200
    Search:
      name: q
      in: query
      required: false
      description: Search only endpoint-declared fields; never expands authorization scope.
      schema:
        type: string
        minLength: 1
        maxLength: 100
    ClassSectionFilter:
      name: classSectionId
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/OpaqueId'
...
    ClassSectionId:
      name: classSectionId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/OpaqueId' }
...
    RosterImportId:
      name: rosterImportId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/OpaqueId' }
    AlignmentResultId:
      name: alignmentResultId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/OpaqueId' }
...
  requestBodies:
    VersionedReasonBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/VersionedReasonRequest'
...
  responses:
...
    RosterImportSuccess:
      description: Roster import version
      content:
        application/json:
          schema: { $ref: '#/components/schemas/RosterImportEnvelope' }
    RosterImportListSuccess:
      description: Paged roster import versions
      content:
        application/json:
          schema: { $ref: '#/components/schemas/RosterImportListEnvelope' }
    RosterEntryListSuccess:
      description: Paged roster entries
      content:
        application/json:
          schema: { $ref: '#/components/schemas/RosterEntryListEnvelope' }
    AlignmentRunSuccess:
      description: Alignment run accepted; results are read from the result collection
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AlignmentRunEnvelope' }
    AlignmentResultSuccess:
      description: Roster alignment result
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AlignmentResultEnvelope' }
    AlignmentResultListSuccess:
      description: Paged roster alignment results
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AlignmentResultListEnvelope' }
...
    BadRequest:
      $ref: '#/components/responses/Error400'
    Unauthorized:
      $ref: '#/components/responses/Error401'
    Forbidden:
      $ref: '#/components/responses/Error403'
    NotFound:
      $ref: '#/components/responses/Error404'
...
    Conflict:
      $ref: '#/components/responses/Error409'
    UnprocessableEntity:
      $ref: '#/components/responses/Error422'
...
    Error400:
      description: Bad request
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error401:
      description: Unauthorized
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error403:
      description: Forbidden
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error404:
      description: Not found
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error409:
      description: Conflict
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Error422:
      description: Unprocessable entity
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
    OpaqueId:
      type: string
      minLength: 1
      maxLength: 64
      pattern: '^[A-Za-z0-9._:-]+$'
      description: Opaque internal ID; clients must not parse prefixes or ordering.
    Sha256:
      type: string
      pattern: '^[a-f0-9]{64}$'
...
    PositiveVersion:
      type: integer
      format: int64
      minimum: 1
...
    RosterImportSource:
      type: string
      enum: [FILE, OFFICIAL_API]
    RosterFileFormat:
      type: string
      enum: [CSV]
    RosterImportStatus:
      type: string
      enum: [RECEIVED, VALIDATING, VALIDATED, FAILED]
    RosterRowValidationStatus:
      type: string
      enum: [VALID, INVALID, DUPLICATED]
    RosterAlignmentStatus:
      type: string
      enum: [MATCHED, MISSING_IN_PLATFORM, EXTRA_IN_PLATFORM, WRONG_COURSE, IDENTITY_CONFLICT, DUPLICATED]
    RosterResolutionStatus:
      type: string
      enum: [PENDING, CONFIRMED, RESOLVED, IGNORED]
...
    RosterResolutionEvidenceType:
      type: string
      enum: [NEW_ALIGNMENT_RESULT, ENROLLMENT_STATUS_EVENT, OFFICIAL_ROSTER_VERSION]
...
    VersionedReasonRequest:
      type: object
      additionalProperties: false
      required: [reason, expectedVersion]
      properties:
        reason: { type: string, minLength: 1, maxLength: 1000 }
        expectedVersion: { $ref: '#/components/schemas/PositiveVersion' }
...
    RosterFieldMapping:
      type: object
      additionalProperties: false
      required: [studentNumber, fullName]
      properties:
        studentNumber: { type: string, minLength: 1, maxLength: 128 }
        fullName: { type: string, minLength: 1, maxLength: 128 }
        gender: { type: [string, 'null'], maxLength: 128 }
        gradeYear: { type: [string, 'null'], maxLength: 128 }
        collegeName: { type: [string, 'null'], maxLength: 128 }
        majorName: { type: [string, 'null'], maxLength: 128 }
        administrativeClassName: { type: [string, 'null'], maxLength: 128 }
      description: Canonical field to exact UTF-8 CSV header mapping. Values must be distinct after trim and contain no control characters.
    CreateRosterImportRequest:
      type: object
      additionalProperties: false
      required: [source, fileFormat, file, fieldMappingSnapshot]
      properties:
        source: { $ref: '#/components/schemas/RosterImportSource' }
        fileFormat: { $ref: '#/components/schemas/RosterFileFormat' }
        file: { type: string, format: binary, contentMediaType: text/csv, description: Streamed UTF-8 CSV source file. The server derives and sanitizes the display filename from multipart metadata. }
        fileChecksumSha256: { oneOf: [{ $ref: '#/components/schemas/Sha256' }, { type: 'null' }] }
        fieldMappingSnapshot: { $ref: '#/components/schemas/RosterFieldMapping' }
      description: V1 accepts only source=FILE with fileFormat=CSV. OFFICIAL_API is a reserved enum branch and returns ROSTER_IMPORT_SOURCE_UNSUPPORTED with no Import side effect until a trusted connector is approved.
    RollbackRosterImportRequest:
      type: object
      additionalProperties: false
      required: [expectedCurrentRosterImportId, expectedVersion, reason]
      properties:
        expectedCurrentRosterImportId: { $ref: '#/components/schemas/OpaqueId' }
        expectedVersion:
          $ref: '#/components/schemas/PositiveVersion'
          description: Optimistic version of expectedCurrentRosterImportId, not the target versionNumber.
        reason: { type: string, minLength: 1, maxLength: 1000 }
    RunAlignmentRequest:
      type: object
      additionalProperties: false
      required: [expectedRosterImportVersion]
      properties:
        expectedRosterImportVersion: { $ref: '#/components/schemas/PositiveVersion' }
      description: The server, never the client, freezes the ACTIVE Enrollment snapshot and computes its SHA-256 fingerprint.
    ResolveAlignmentRequest:
      type: object
      additionalProperties: false
      required: [resolutionNote, evidenceType, evidenceReferenceId, expectedVersion]
      properties:
        resolutionNote: { type: string, minLength: 1, maxLength: 1000 }
        evidenceType: { $ref: '#/components/schemas/RosterResolutionEvidenceType' }
        evidenceReferenceId: { $ref: '#/components/schemas/OpaqueId' }
        expectedVersion: { $ref: '#/components/schemas/PositiveVersion' }
...
    RosterImportEnvelope: { $ref: '#/components/schemas/RosterImportResponse' }
    RosterImportListEnvelope: { $ref: '#/components/schemas/RosterImportListResponse' }
    RosterEntryListEnvelope: { $ref: '#/components/schemas/RosterEntryListResponse' }
    AlignmentRunEnvelope: { $ref: '#/components/schemas/AlignmentRunResponse' }
    AlignmentResultEnvelope: { $ref: '#/components/schemas/AlignmentResultResponse' }
    AlignmentResultListEnvelope: { $ref: '#/components/schemas/AlignmentResultListResponse' }
...
    RosterImportListResponse:
      type: object
      required: [data, meta]
      properties: { data: { type: array, items: { $ref: '#/components/schemas/OfficialRosterImport' } }, meta: { $ref: '#/components/schemas/PagedMeta' } }
...